move smartagentextension to its own module
