+++ deployments/ansible/roles/collector/README.md
@@ -90,7 +90,7 @@ $> ansible-playbook playbook.yaml -e start_service=false
   that nothing will be copied and existing `splunk_otel_collector_config` will be used)
 
 - `splunk_bundle_dir`: The path to the [Smart Agent bundle directory](
-  https://github.com/signalfx/splunk-otel-collector/blob/main/internal/extension/smartagentextension/README.md).
+  https://github.com/signalfx/splunk-otel-collector/blob/main/pkg/extension/smartagentextension/README.md).
   The default path is provided by the collector package. If the specified path
   is changed from the default value, the path should be an existing directory
   on the node. The `SPLUNK_BUNDLE_DIR` environment variable will be set to
