+++ pkg/extension/smartagentextension/README.md
