+++ pkg/extension/smartagentextension/config.go
