+++ pkg/extension/smartagentextension/config_linux_test.go
