+++ pkg/extension/smartagentextension/config_test.go
