+++ pkg/extension/smartagentextension/config_windows_test.go
