+++ pkg/extension/smartagentextension/extension.go
