+++ pkg/extension/smartagentextension/factory.go
