+++ pkg/extension/smartagentextension/factory_test.go
