+++ pkg/extension/smartagentextension/testdata/config.yaml
