+++ pkg/extension/smartagentextension/testdata/invalid_config.yaml
