+++ pkg/receiver/smartagent/receiver.go
@@ -37,7 +37,7 @@ import (
 	"go.opentelemetry.io/collector/consumer"
 	"go.uber.org/zap"
 
-	"github.com/signalfx/splunk-otel-collector/internal/extension/smartagentextension"
+	"github.com/signalfx/splunk-otel-collector/extension/smartagentextension"
 )
 
 const setOutputErrMsg = "unable to set output field of monitor"
